+++ Backend/src/controllers/players.ts
@@ -1,6 +1,7 @@
 import { Router } from "express";
 import { authenticateToken } from "../utils/middleware";
 import playerService from "../services/playerService";
+import seasonService from "../services/seasonService";
 
 
 const playersRouter = Router();
@@ -9,7 +10,14 @@ const playersRouter = Router();
 playersRouter.get('/season/:id', async (req, res) => {
   const { id } = req.params;
   try {
-    const players = await playerService.getAllPlayersStatsBySeason(parseInt(id));
+    const id_number = parseInt(id);
+    const season = await seasonService.getSeasonById(id_number);
+    if (!season) {
+      res.status(404).json({ error: 'season not found' });
+      return;
+    }
+
+    const players = await playerService.getAllPlayersStatsBySeason(id_number);
     if (!players) {
       res.status(404).json({ error: 'players not found' });
       return;
@@ -73,6 +81,12 @@ playersRouter.delete('/:id', authenticateToken, async (req, res) => {
       res.status(400).json({ error: 'malformatted id' });
       return;
     }
+
+    const player = await playerService.getPlayerById(id_number);
+    if (!player) {
+      res.status(404).json({ error: 'player not found' });
+      return;
+    }
     await playerService.deletePlayer(id_number);
 
     res.status(204).end();
@@ -92,9 +106,16 @@ playersRouter.put('/:id', authenticateToken, async (req, res) => {
   }
 
   try {
-    const player = await playerService.updatePlayer(parseInt(id), name, nickname, number);
+    const id_number = parseInt(id);
 
-    res.json(player);
+    if (!(await playerService.getPlayerById(id_number))) {
+      res.status(404).json({ error: 'player not found' });
+      return;
+    }
+
+    const player = await playerService.updatePlayer(id_number, name, nickname, number);
+
+    res.status(204).json(player);
   } catch (error) {
     res.status(500).json({ error: 'Something went wrong' });
     console.log(error);
